+++ pkgtools/pkgtasks/files/t/t_postinstall.sh
@@ -47,54 +47,93 @@ test_setup()
 }
 
 # Mock actions that return the truthiness of environment variables.
+mock_pkgtask()
+{
+	# required parameters
+	local name="$1"; shift
+	local varname="$1"; shift
+
+	# pkgtask parameters
+	local arg
+	local silent=
+	local OPTIND=1
+	while getopts ":s" arg "$@"; do
+		case $arg in
+		s)	silent="yes" ;;
+		*)	return 127 ;;
+		esac
+	done
+	shift $(( ${OPTIND} - 1 ))
+	local action="$1"; shift
+	local stage="$1"; shift
+
+	local value symbol result
+	eval value="\${$varname}"
+	if [ "$value" = "yes" ]; then
+		symbol=">"
+		result=0
+	else
+		symbol="!"
+		result=1
+	fi
+
+	# Skip output if doing a "check" action.
+	case $action in
+	check-*)
+		: "do nothing" ;;
+	*)	[ -n "$silent" ] || echo "$symbol $name" ;;
+	esac
+	return $result
+}
+
 task_directories()
 {
-	[ "${TASK_DIRECTORIES_SUCCESS}" = "yes" ]
+	mock_pkgtask "directories" "TASK_DIRECTORIES_SUCCESS" "$@"
 }
 
 task_files()
 {
-	[ "${TASK_FILES_SUCCESS}" = "yes" ]
+	mock_pkgtask "files" "TASK_FILES_SUCCESS" "$@"
 }
 
 task_fonts()
 {
-	[ "${TASK_FONTS_SUCCESS}" = "yes" ]
+	mock_pkgtask "fonts" "TASK_FONTS_SUCCESS" "$@"
 }
 
 task_function()
 {
-	[ "${TASK_FUNCTION_SUCCESS}" = "yes" ]
+	mock_pkgtask "function" "TASK_FUNCTION_SUCCESS" "$@"
 }
 
 task_icon_themes()
 {
-	[ "${TASK_ICON_THEMES_SUCCESS}" = "yes" ]
+	mock_pkgtask "icon_themes" "TASK_ICON_THEMES_SUCCESS" "$@"
 }
 
 task_info_files()
 {
-	[ "${TASK_INFO_FILES_SUCCESS}" = "yes" ]
+	mock_pkgtask "info_files" "TASK_INFO_FILES_SUCCESS" "$@"
 }
 
 task_ocaml_findlib()
 {
-	[ "${TASK_OCAML_FINDLIB_SUCCESS}" = "yes" ]
+	mock_pkgtask "ocaml_findlib" "TASK_OCAML_FINDLIB_SUCCESS" "$@"
 }
 
 task_permissions()
 {
-	[ "${TASK_PERMISSIONS_SUCCESS}" = "yes" ]
+	mock_pkgtask "permissions" "TASK_PERMISSIONS_SUCCESS" "$@"
 }
 
 task_shells()
 {
-	[ "${TASK_SHELLS_SUCCESS}" = "yes" ]
+	mock_pkgtask "shells" "TASK_SHELLS_SUCCESS" "$@"
 }
 
 task_shlibs()
 {
-	[ "${TASK_SHLIBS_SUCCESS}" = "yes" ]
+	mock_pkgtask "shlibs" "TASK_SHLIBS_SUCCESS" "$@"
 }
 
 # Only succeed if all of the actions were successful.
@@ -209,6 +248,44 @@ test11()
 	else
 		return 1
 	fi
+	return 0
+}
+
+test12()
+{
+	: ${CAT:=cat}
+
+	describe="\${TASK_VERBOSE} = none"
+	TASK_VERBOSE=none
+	task_postinstall "$datafile" > output
+	${CAT} output
+	if [ -s "$output" ]; then
+		# no output expected
+		return 1
+	fi
+	return 0
+}
+
+test13()
+{
+	: ${CAT:=cat}
+	: ${GREP:=grep}
+
+	describe="\${TASK_VERBOSE} = all"
+	TASK_VERBOSE=all
+	task_postinstall "$datafile" > output
+	${CAT} output
+	for task in \
+		files fonts function icon_themes info_files \
+		ocaml_findlib permissions shells shlibs
+	do
+		if ${GREP} -q "$task" output; then
+			: "success"
+		else
+			describe="$describe: '$task' missing!"
+			return 1
+		fi
+	done
 	return 0
 }